+++ types/util.go
@@ -8,6 +8,8 @@ import (
 	cmn "github.com/tendermint/tmlibs/common"
 )
 
+//------------------------------------------------------------------------------
+
 // Validators is a list of validators that implements the Sort interface
 type Validators []*Validator
 
@@ -26,13 +28,6 @@ func (v Validators) Swap(i, j int) {
 	v[j] = v1
 }
 
-//-------------------------------------
-
-type validatorPretty struct {
-	PubKey data.Bytes `json:"pub_key"`
-	Power  int64      `json:"power"`
-}
-
 func ValidatorsString(vs Validators) string {
 	s := make([]validatorPretty, len(vs))
 	for i, v := range vs {
@@ -43,4 +38,29 @@ func ValidatorsString(vs Validators) string {
 		cmn.PanicSanity(err.Error())
 	}
 	return string(b)
+}
+
+type validatorPretty struct {
+	PubKey data.Bytes `json:"pub_key"`
+	Power  int64      `json:"power"`
+}
+
+//------------------------------------------------------------------------------
+
+// KVPairInt is a helper method to build KV pair with an integer value.
+func KVPairInt(key string, val int64) *KVPair {
+	return &KVPair{
+		Key:       key,
+		ValueInt:  val,
+		ValueType: KVPair_INT,
+	}
+}
+
+// KVPairString is a helper method to build KV pair with a string value.
+func KVPairString(key, val string) *KVPair {
+	return &KVPair{
+		Key:         key,
+		ValueString: val,
+		ValueType:   KVPair_STRING,
+	}
 }